requirements.txt
discord.py[voice]
jishaku
pomice
humanize
async-cse